move: rename module to spyder_remote_services
