+++ spyder_remote_services/jupyter_client/manager.py
